+++ tests/integrated/test-misc.R
@@ -1,19 +1,19 @@
 test_that("audit_logs work", {
-  logs <- test_conn_1$audit_logs()
+  logs <- client$audit_logs()
   expect_gt(length(logs$results), 0)
 
-  logs2 <- test_conn_1$audit_logs(nxt = logs$paging$cursors$`next`)
+  logs2 <- client$audit_logs(nxt = logs$paging$cursors$`next`)
   expect_gt(length(logs2$results), 0)
 })
 
 test_that("server_settings work", {
-  ss <- test_conn_1$server_settings()
+  ss <- client$server_settings()
 
   expect_gt(length(ss), 0)
 })
 
 test_that("server_settings_r work", {
-  ssr <- test_conn_1$server_settings_r()
+  ssr <- client$server_settings_r()
 
   expect_gt(length(ssr$installations), 0)
 })