+++ tests/integrated/test-tags.R
@@ -24,22 +24,22 @@ check_tag_exists <- function(con, id) {
 # Tests ---------------------------------------
 
 test_that("create tags works", {
-  parent_tag <<- test_conn_1$tag_create(parent_tag_name)
-  child_tag <<- test_conn_1$tag_create(child_tag_name, parent_tag$id)
+  parent_tag <<- client$tag_create(parent_tag_name)
+  child_tag <<- client$tag_create(child_tag_name, parent_tag$id)
 
   expect_identical(parent_tag$name, parent_tag_name)
   expect_identical(child_tag$name, child_tag_name)
   expect_identical(child_tag$parent_id, parent_tag$id)
 
   # tag_id lookup works
-  test_conn_1$get_tags(TRUE)
-  expect_equal(child_tag$id, test_conn_1$get_tag_id(child_tag_name))
+  client$get_tags(TRUE)
+  expect_equal(child_tag$id, client$get_tag_id(child_tag_name))
 })
 
 ## Test high level functions --------------------------------------------------
 
 test_that("get_tags works", {
-  atags <- get_tags(test_conn_1)
+  atags <- get_tags(client)
   expect_s3_class(atags, "connect_tag_tree")
 })
 
@@ -49,50 +49,50 @@ test_that("create_tag and delete_tag works", {
   ctag_2 <- uuid::UUIDgenerate(use.time = TRUE)
   ctag_3 <- uuid::UUIDgenerate(use.time = TRUE)
 
-  capture.output(res <- create_tag(test_conn_1, ptag_1))
+  capture.output(res <- create_tag(client, ptag_1))
   expect_true(validate_R6_class(res, "Connect"))
 
-  tags <- get_tags(test_conn_1)
-  capture.output(create_tag(test_conn_1, ctag_1, tags[[ptag_1]]))
+  tags <- get_tags(client)
+  capture.output(create_tag(client, ctag_1, tags[[ptag_1]]))
 
-  tags <- get_tags(test_conn_1)
-  capture.output(create_tag(test_conn_1, ctag_2, tags[[ptag_1]][[ctag_1]]))
+  tags <- get_tags(client)
+  capture.output(create_tag(client, ctag_2, tags[[ptag_1]][[ctag_1]]))
 
-  tags <- get_tags(test_conn_1)
+  tags <- get_tags(client)
   capture.output(create_tag(
-    test_conn_1,
+    client,
     ctag_3,
     tags[[ptag_1]][[ctag_1]][[ctag_2]]
   ))
 
-  tags <- get_tags(test_conn_1)
+  tags <- get_tags(client)
 
-  delete_tag(test_conn_1, tags[[ptag_1]][[ctag_1]][[ctag_2]][[ctag_3]])
+  delete_tag(client, tags[[ptag_1]][[ctag_1]][[ctag_2]][[ctag_3]])
   expect_false(check_tag_exists(
-    test_conn_1,
+    client,
     tags[[ptag_1]][[ctag_1]][[ctag_2]][[ctag_3]][["id"]]
   ))
 
-  delete_tag(test_conn_1, tags[[ptag_1]][[ctag_1]][[ctag_2]])
+  delete_tag(client, tags[[ptag_1]][[ctag_1]][[ctag_2]])
   expect_false(check_tag_exists(
-    test_conn_1,
+    client,
     tags[[ptag_1]][[ctag_1]][[ctag_2]][["id"]]
   ))
 
-  delete_tag(test_conn_1, tags[[ptag_1]][[ctag_1]])
-  expect_false(check_tag_exists(test_conn_1, tags[[ptag_1]][[ctag_1]][["id"]]))
+  delete_tag(client, tags[[ptag_1]][[ctag_1]])
+  expect_false(check_tag_exists(client, tags[[ptag_1]][[ctag_1]][["id"]]))
 
-  res <- delete_tag(test_conn_1, tags[[ptag_1]])
-  expect_false(check_tag_exists(test_conn_1, tags[[ptag_1]][["id"]]))
+  res <- delete_tag(client, tags[[ptag_1]])
+  expect_false(check_tag_exists(client, tags[[ptag_1]][["id"]]))
 
   expect_true(validate_R6_class(res, "Connect"))
 })
 
 test_that("delete_tag errs for whole tree", {
-  alltags <- get_tags(test_conn_1)
+  alltags <- get_tags(client)
 
   expect_error(
-    delete_tag(test_conn_1, alltags),
+    delete_tag(client, alltags),
     "entire tag tree"
   )
 })
@@ -100,10 +100,10 @@ test_that("delete_tag errs for whole tree", {
 test_that("con$tag with id returns just one record", {
   ptag_1 <- uuid::UUIDgenerate(use.time = TRUE)
   ctag_1 <- uuid::UUIDgenerate(use.time = TRUE)
-  capture.output(a1 <- create_tag_tree(test_conn_1, ptag_1, ctag_1))
+  capture.output(a1 <- create_tag_tree(client, ptag_1, ctag_1))
 
-  tags1 <- get_tags(test_conn_1)
-  res <- test_conn_1$tag(tags1[[ptag_1]][[ctag_1]][["id"]])
+  tags1 <- get_tags(client)
+  res <- client$tag(tags1[[ptag_1]][[ctag_1]][["id"]])
   expect_equal(
     names(res),
     c("id", "name", "parent_id", "created_time", "updated_time")
@@ -116,23 +116,23 @@ test_that("create_tag_tree works", {
   ctag_2 <- uuid::UUIDgenerate(use.time = TRUE)
   ctag_3 <- uuid::UUIDgenerate(use.time = TRUE)
 
-  capture.output(a1 <- create_tag_tree(test_conn_1, ptag_1, ctag_1))
+  capture.output(a1 <- create_tag_tree(client, ptag_1, ctag_1))
   expect_true(validate_R6_class(a1, "Connect"))
 
-  tags1 <- get_tags(test_conn_1)
+  tags1 <- get_tags(client)
 
   capture.output(
-    a2 <- create_tag_tree(test_conn_1, ptag_1, ctag_1, ctag_2, ctag_3)
+    a2 <- create_tag_tree(client, ptag_1, ctag_1, ctag_2, ctag_3)
   )
   expect_true(validate_R6_class(a2, "Connect"))
 
-  tags2 <- get_tags(test_conn_1)
+  tags2 <- get_tags(client)
 
   expect_identical(tags1[[ptag_1]][["id"]], tags2[[ptag_1]][["id"]])
 
-  delete_tag(test_conn_1, tags2[[ptag_1]])
+  delete_tag(client, tags2[[ptag_1]])
   expect_error(
-    suppressMessages(test_conn_1$tag(tags2[[ptag_1]][["id"]])),
+    suppressMessages(client$tag(tags2[[ptag_1]][["id"]])),
     "Not Found"
   )
 })
@@ -145,7 +145,7 @@ test_that("get_content_tags and set_content_tags works", {
   ctag_2_1 <- uuid::UUIDgenerate(use.time = TRUE)
 
   app1 <- deploy(
-    test_conn_1,
+    client,
     bundle_dir(rprojroot::find_package_root_file(
       "tests",
       "testthat",
@@ -155,9 +155,9 @@ test_that("get_content_tags and set_content_tags works", {
   )
 
   capture.output(
-    tmp1 <- create_tag_tree(test_conn_1, ptag_1, ctag_1_1, ctag_1_2)
+    tmp1 <- create_tag_tree(client, ptag_1, ctag_1_1, ctag_1_2)
   )
-  capture.output(tmp2 <- create_tag_tree(test_conn_1, ptag_1, ctag_2_1))
+  capture.output(tmp2 <- create_tag_tree(client, ptag_1, ctag_2_1))
 
   expect_true(validate_R6_class(tmp1, "Connect"))
   expect_true(validate_R6_class(tmp2, "Connect"))
@@ -165,7 +165,7 @@ test_that("get_content_tags and set_content_tags works", {
   ct <- get_content_tags(app1)
   expect_length(ct, 0)
 
-  all_tags <- get_tags(test_conn_1)
+  all_tags <- get_tags(client)
 
   c1o <- capture.output(
     c1 <- set_content_tags(
@@ -198,7 +198,7 @@ test_that("set_content_tag_tree works", {
   ctag_2_1 <- uuid::UUIDgenerate(use.time = TRUE)
 
   app1 <- deploy(
-    test_conn_1,
+    client,
     bundle_dir(rprojroot::find_package_root_file(
       "tests",
       "testthat",
@@ -208,9 +208,9 @@ test_that("set_content_tag_tree works", {
   )
 
   capture.output(
-    tmp1 <- create_tag_tree(test_conn_1, ptag_1, ctag_1_1, ctag_1_2)
+    tmp1 <- create_tag_tree(client, ptag_1, ctag_1_1, ctag_1_2)
   )
-  capture.output(tmp2 <- create_tag_tree(test_conn_1, ptag_1, ctag_2_1))
+  capture.output(tmp2 <- create_tag_tree(client, ptag_1, ctag_2_1))
 
   expect_true(validate_R6_class(tmp1, "Connect"))
   expect_true(validate_R6_class(tmp2, "Connect"))